refactor: 길이 검증 로직 수정

+++ src/main/java/com/programmers/week/item/domain/Car.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Car extends Item {
 
-    private static final int MIN_POWER = 0;
+    private static final long MIN_POWER = 0;
     private static final long MAX_POWER = 1000000;
     private long power;
